将count字段更改为double类型

+++ app/Http/Requests/Material/MaterialShowRequest.php
@@ -29,8 +29,8 @@ class MaterialShowRequest extends FormRequest
             'number' => 'string',
             'title' => 'string',
             'unit' => 'string',
-            'countBegin' => 'integer',
-            'countEnd' => 'integer',
+            'countBegin' => 'Numeric',
+            'countEnd' => 'Numeric',
             'sortField' => 'string',
             'sortord' => 'in:desc,asc'
         ];
