+++ app/Http/Requests/Product/ProductShowRequest.php
@@ -29,8 +29,8 @@ class ProductShowRequest extends FormRequest
             'number' => 'string',
             'title' => 'string',
             'unit' => 'string',
-            'countBegin' => 'integer',
-            'countEnd' => 'integer',
+            'countBegin' => 'Numeric',
+            'countEnd' => 'Numeric',
             'sortField' => 'string',
             'sortord' => 'in:desc,asc'
         ];
